requirements.txt
python-dotenv==1.0.1
requests==2.31.0
pyodbc==4.0.34
psycopg2==2.9.9
tenacity==8.0.1
numpy==1.23.5
pandas==1.5.3
sqlalchemy==1.4.39
tqdm==4.64.1